Revert "[team-18][FE] Sonya"
