+++ FE/tracker/src/components/issueList/issueList.tsx
@@ -1,61 +1,280 @@
 import * as React from "react";
-import axios from "axios";
-
-// MUI Material
+import { alpha } from "@mui/material/styles";
 import Box from "@mui/material/Box";
 import Table from "@mui/material/Table";
 import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
 import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
+import TableSortLabel from "@mui/material/TableSortLabel";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
 import Paper from "@mui/material/Paper";
 import Checkbox from "@mui/material/Checkbox";
+import IconButton from "@mui/material/IconButton";
+import Tooltip from "@mui/material/Tooltip";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Switch from "@mui/material/Switch";
+import DeleteIcon from "@mui/icons-material/Delete";
+import FilterListIcon from "@mui/icons-material/FilterList";
+import { visuallyHidden } from "@mui/utils";
 
-// MUI Theme
-// import { ThemeProvider } from "@mui/material/styles";
-import { listTheme } from "../../mui-style/muiTheme";
+interface Data {
+    calories: number;
+    carbs: number;
+    fat: number;
+    name: string;
+    protein: number;
+}
 
-// Types
-import { Data, Order } from "./IssueLIst.types";
+function createData(
+    name: string,
+    calories: number,
+    fat: number,
+    carbs: number,
+    protein: number
+): Data {
+    return {
+        name,
+        calories,
+        fat,
+        carbs,
+        protein,
+    };
+}
 
-// Functions
-import {
-    descendingComparator,
-    getComparator,
-    stableSort,
-    createData,
-} from "./IssueList.function.tsx";
+const rows = [
+    createData("Cupcake", 305, 3.7, 67, 4.3),
+    createData("Donut", 452, 25.0, 51, 4.9),
+    createData("Eclair", 262, 16.0, 24, 6.0),
+    createData("Frozen yoghurt", 159, 6.0, 24, 4.0),
+    createData("Gingerbread", 356, 16.0, 49, 3.9),
+    createData("Honeycomb", 408, 3.2, 87, 6.5),
+    createData("Ice cream sandwich", 237, 9.0, 37, 4.3),
+    createData("Jelly Bean", 375, 0.0, 94, 0.0),
+    createData("KitKat", 518, 26.0, 65, 7.0),
+    createData("Lollipop", 392, 0.2, 98, 0.0),
+    createData("Marshmallow", 318, 0, 81, 2.0),
+    createData("Nougat", 360, 19.0, 9, 37.0),
+    createData("Oreo", 437, 18.0, 63, 4.0),
+];
 
-// Table Parts
-import EnhancedTableHead from "./issueListHead.tsx";
-import { EnhancedTableToolbar } from "./issueListToolbar.tsx";
+function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
+    if (b[orderBy] < a[orderBy]) {
+        return -1;
+    }
+    if (b[orderBy] > a[orderBy]) {
+        return 1;
+    }
+    return 0;
+}
 
-// const rows = [
-//     createData("Cupcake", 305, 3.7, 67, 4.3),
-//     createData("Donut", 452, 25.0, 51, 4.9),
-//     createData("Eclair", 262, 16.0, 24, 6.0),
-//     createData("Frozen yoghurt", 159, 6.0, 24, 4.0),
-//     createData("Gingerbread", 356, 16.0, 49, 3.9),
-//     createData("Honeycomb", 408, 3.2, 87, 6.5),
-//     createData("Ice cream sandwich", 237, 9.0, 37, 4.3),
-//     createData("Jelly Bean", 375, 0.0, 94, 0.0),
-//     createData("KitKat", 518, 26.0, 65, 7.0),
-//     createData("Lollipop", 392, 0.2, 98, 0.0),
-//     createData("Marshmallow", 318, 0, 81, 2.0),
-//     createData("Nougat", 360, 19.0, 9, 37.0),
-//     createData("Oreo", 437, 18.0, 63, 4.0),
-// ];
+type Order = "asc" | "desc";
+
+function getComparator<Key extends keyof any>(
+    order: Order,
+    orderBy: Key
+): (
+    a: { [key in Key]: number | string },
+    b: { [key in Key]: number | string }
+) => number {
+    return order === "desc"
+        ? (a, b) => descendingComparator(a, b, orderBy)
+        : (a, b) => -descendingComparator(a, b, orderBy);
+}
+
+// This method is created for cross-browser compatibility, if you don't
+// need to support IE11, you can use Array.prototype.sort() directly
+function stableSort<T>(
+    array: readonly T[],
+    comparator: (a: T, b: T) => number
+) {
+    const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
+    stabilizedThis.sort((a, b) => {
+        const order = comparator(a[0], b[0]);
+        if (order !== 0) {
+            return order;
+        }
+        return a[1] - b[1];
+    });
+    return stabilizedThis.map((el) => el[0]);
+}
+
+interface HeadCell {
+    disablePadding: boolean;
+    id: keyof Data;
+    label: string;
+    numeric: boolean;
+}
+
+const headCells: readonly HeadCell[] = [
+    {
+        id: "name",
+        numeric: false,
+        disablePadding: true,
+        label: "Dessert (100g serving)",
+    },
+    {
+        id: "calories",
+        numeric: true,
+        disablePadding: false,
+        label: "Calories",
+    },
+    {
+        id: "fat",
+        numeric: true,
+        disablePadding: false,
+        label: "Fat (g)",
+    },
+    {
+        id: "carbs",
+        numeric: true,
+        disablePadding: false,
+        label: "Carbs (g)",
+    },
+    {
+        id: "protein",
+        numeric: true,
+        disablePadding: false,
+        label: "Protein (g)",
+    },
+];
+
+interface EnhancedTableProps {
+    numSelected: number;
+    onRequestSort: (
+        event: React.MouseEvent<unknown>,
+        property: keyof Data
+    ) => void;
+    onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
+    order: Order;
+    orderBy: string;
+    rowCount: number;
+}
+
+function EnhancedTableHead(props: EnhancedTableProps) {
+    const {
+        onSelectAllClick,
+        order,
+        orderBy,
+        numSelected,
+        rowCount,
+        onRequestSort,
+    } = props;
+    const createSortHandler =
+        (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
+            onRequestSort(event, property);
+        };
+
+    return (
+        <TableHead>
+            <TableRow>
+                <TableCell padding="checkbox">
+                    <Checkbox
+                        color="primary"
+                        indeterminate={
+                            numSelected > 0 && numSelected < rowCount
+                        }
+                        checked={rowCount > 0 && numSelected === rowCount}
+                        onChange={onSelectAllClick}
+                        inputProps={{
+                            "aria-label": "select all desserts",
+                        }}
+                    />
+                </TableCell>
+                {headCells.map((headCell) => (
+                    <TableCell
+                        key={headCell.id}
+                        align={headCell.numeric ? "right" : "left"}
+                        padding={headCell.disablePadding ? "none" : "normal"}
+                        sortDirection={orderBy === headCell.id ? order : false}
+                    >
+                        <TableSortLabel
+                            active={orderBy === headCell.id}
+                            direction={orderBy === headCell.id ? order : "asc"}
+                            onClick={createSortHandler(headCell.id)}
+                        >
+                            {headCell.label}
+                            {orderBy === headCell.id ? (
+                                <Box component="span" sx={visuallyHidden}>
+                                    {order === "desc"
+                                        ? "sorted descending"
+                                        : "sorted ascending"}
+                                </Box>
+                            ) : null}
+                        </TableSortLabel>
+                    </TableCell>
+                ))}
+            </TableRow>
+        </TableHead>
+    );
+}
+
+interface EnhancedTableToolbarProps {
+    numSelected: number;
+}
+
+const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
+    const { numSelected } = props;
+
+    return (
+        <Toolbar
+            sx={{
+                pl: { sm: 2 },
+                pr: { xs: 1, sm: 1 },
+                ...(numSelected > 0 && {
+                    bgcolor: (theme) =>
+                        alpha(
+                            theme.palette.primary.main,
+                            theme.palette.action.activatedOpacity
+                        ),
+                }),
+            }}
+        >
+            {numSelected > 0 ? (
+                <Typography
+                    sx={{ flex: "1 1 100%" }}
+                    color="inherit"
+                    variant="subtitle1"
+                    component="div"
+                >
+                    {numSelected} selected
+                </Typography>
+            ) : (
+                <Typography
+                    sx={{ flex: "1 1 100%" }}
+                    variant="h6"
+                    id="tableTitle"
+                    component="div"
+                >
+                    Nutrition
+                </Typography>
+            )}
+            {numSelected > 0 ? (
+                <Tooltip title="Delete">
+                    <IconButton>
+                        <DeleteIcon />
+                    </IconButton>
+                </Tooltip>
+            ) : (
+                <Tooltip title="Filter list">
+                    <IconButton>
+                        <FilterListIcon />
+                    </IconButton>
+                </Tooltip>
+            )}
+        </Toolbar>
+    );
+};
 
 export default function EnhancedTable() {
     const [order, setOrder] = React.useState<Order>("asc");
-    const [orderBy, setOrderBy] = React.useState<keyof Data>("writer");
+    const [orderBy, setOrderBy] = React.useState<keyof Data>("calories");
     const [selected, setSelected] = React.useState<readonly string[]>([]);
     const [page, setPage] = React.useState(0);
     const [dense, setDense] = React.useState(false);
-    const [rows, setRows] = React.useState([]);
     const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
     const handleRequestSort = (
@@ -70,14 +289,12 @@ export default function EnhancedTable() {
     const handleSelectAllClick = (
         event: React.ChangeEvent<HTMLInputElement>
     ) => {
-        // if (event.target.checked) {
-        //     const newSelecteds = rows.map((n) => n.name);
-        //     setSelected(newSelecteds);
-        //     return;
-        // }
-        // setSelected([]);
-
-        console.log("handleSelectAllClick");
+        if (event.target.checked) {
+            const newSelecteds = rows.map((n) => n.name);
+            setSelected(newSelecteds);
+            return;
+        }
+        setSelected([]);
     };
 
     const handleClick = (event: React.MouseEvent<unknown>, name: string) => {
@@ -117,26 +334,12 @@ export default function EnhancedTable() {
 
     const isSelected = (name: string) => selected.indexOf(name) !== -1;
 
-    const fetchData = () => {
-        return axios
-            .get(
-                "https://313952d4-dd03-4472-a26f-e82b8c0038da.mock.pstmn.io/api/issues?page=1&status=OPEN"
-            )
-            .then((res) => console.log(res.data));
-    };
-
-    React.useEffect(() => {
-        // Postman Api 문서에서 데이터 받아오기
-        console.log("데이터중복방지?");
-        fetchData();
-    }, []);
-
     // Avoid a layout jump when reaching the last page with empty rows.
     const emptyRows =
         page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
 
     return (
-        <Box sx={{ width: "800px", mt: 5, marginLeft: 10 }}>
+        <Box sx={{ width: "100%" }}>
             <Paper sx={{ width: "100%", mb: 2 }}>
                 <EnhancedTableToolbar numSelected={selected.length} />
                 <TableContainer>
