--- FE/tracker/src/components/issueList/issueListHead.tsx
@@ -1,97 +1,0 @@
-import * as React from "react";
-import Box from "@mui/material/Box";
-import TableCell from "@mui/material/TableCell";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import TableSortLabel from "@mui/material/TableSortLabel";
-import Checkbox from "@mui/material/Checkbox";
-import { visuallyHidden } from "@mui/utils";
-
-import { Data, HeadCell, EnhancedTableProps } from "./IssueLIst.types";
-
-const headCells: readonly HeadCell[] = [
-    {
-        id: "name",
-        numeric: false,
-        disablePadding: true,
-        label: "이슈(날짜순)",
-    },
-    {
-        id: "issueDate",
-        numeric: false,
-        disablePadding: true,
-        label: "날짜",
-    },
-    {
-        id: "writer",
-        numeric: true,
-        disablePadding: false,
-        label: "작성자",
-    },
-    {
-        id: "reviewer",
-        numeric: true,
-        disablePadding: false,
-        label: "담당자",
-    },
-];
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-    const {
-        onSelectAllClick,
-        order,
-        orderBy,
-        numSelected,
-        rowCount,
-        onRequestSort,
-    } = props;
-    const createSortHandler =
-        (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
-            onRequestSort(event, property);
-        };
-
-    return (
-        <TableHead>
-            <TableRow>
-                <TableCell padding="checkbox">
-                    <Checkbox
-                        color="primary"
-                        indeterminate={
-                            numSelected > 0 && numSelected < rowCount
-                        }
-                        checked={rowCount > 0 && numSelected === rowCount}
-                        onChange={onSelectAllClick}
-                        inputProps={{
-                            "aria-label": "select all desserts",
-                        }}
-                    />
-                </TableCell>
-                {headCells.map((headCell) => (
-                    <TableCell
-                        key={headCell.id}
-                        align={headCell.numeric ? "right" : "left"}
-                        padding={headCell.disablePadding ? "none" : "normal"}
-                        sortDirection={orderBy === headCell.id ? order : false}
-                    >
-                        <TableSortLabel
-                            active={orderBy === headCell.id}
-                            direction={orderBy === headCell.id ? order : "asc"}
-                            onClick={createSortHandler(headCell.id)}
-                        >
-                            {headCell.label}
-                            {orderBy === headCell.id ? (
-                                <Box component="span" sx={visuallyHidden}>
-                                    {order === "desc"
-                                        ? "sorted descending"
-                                        : "sorted ascending"}
-                                </Box>
-                            ) : null}
-                        </TableSortLabel>
-                    </TableCell>
-                ))}
-            </TableRow>
-        </TableHead>
-    );
-}
-
-export default EnhancedTableHead;
