+++ FE/tracker/src/components/signIn/canvas.jsx
@@ -1,0 +1,55 @@
+import React, { useEffect, useRef } from "react";
+
+function Matrix() {
+    // Get the canvas node and the drawing context
+    const canvasRef = useRef(null);
+    const canvas = canvasRef.current;
+    const ctx = canvas.getContext("2d");
+
+    // set the width and height of the canvas
+    const w = (canvas.width = document.body.offsetWidth);
+    const h = (canvas.height = document.body.offsetHeight);
+
+    const cols = Math.floor(w / 20) + 1;
+    const ypos = Array(cols).fill(0);
+
+    // draw a black rectangle of width and height same as that of the canvas
+    ctx.fillStyle = "#000";
+    ctx.fillRect(0, 0, w, h);
+
+    const drawMatrix = () => {
+        // Draw a semitransparent black rectangle on top of previous drawing
+        ctx.fillStyle = "#0001";
+        ctx.fillRect(0, 0, w, h);
+
+        // Set color to green and font to 15pt monospace in the drawing context
+        ctx.fillStyle = "#0f0";
+        ctx.font = "15pt monospace";
+
+        // for each column put a random character at the end
+        ypos.forEach((y, ind) => {
+            // generate a random character
+            const text = String.fromCharCode(Math.random() * 128);
+
+            // x coordinate of the column, y coordinate is already given
+            const x = ind * 20;
+            // render the character at (x, y)
+            ctx.fillText(text, x, y);
+
+            // randomly reset the end of the column if it's at least 100px high
+            if (y > 100 + Math.random() * 10000) ypos[ind] = 0;
+            // otherwise just move the y coordinate for the column 20px down,
+            else ypos[ind] = y + 20;
+        });
+    };
+
+    useEffect(() => {
+        const intervalGoes = setInterval(() => {
+            drawMatrix();
+        }, 50);
+    });
+
+    return <canvas ref={canvasRef} />;
+}
+
+export default Matrix;
