+++ FE/tracker/src/components/signIn/signIn.jsx
@@ -14,14 +14,10 @@ import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import Copyright from "./copyright.jsx";
-import { matrixTheme } from "../../mui-style/muiTheme.jsx";
-
-import Matrix from "./matrix.jsx";
-const loginUri =
-    "https://github.com/login/oauth/authorize?client_id=2f3993bd2491df08d37b";
+import { matrixTheme } from "../../mui-style/theme.jsx";
 
 export default function SignIn(props) {
-    // const { req: Request, res: Response, next: NextFunction } = props;
+    const { req: Request, res: Response, next: NextFunction } = props;
     // const code = req.query.code;
     // const pathUrl = req.query.state;
 
@@ -47,86 +43,47 @@ export default function SignIn(props) {
     //   return res.redirect(`${config.get<string>('origin')}/oauth/error`);
     // }
 
+    const loginUri =
+        "https://github.com/login/oauth/authorize?client_id=2f3993bd2491df08d37b";
+
     const handleGithubLogin = (event) => {
         event.preventDefault();
         console.log("success");
     };
 
     return (
-        <>
-            <Matrix>
-                <p>hi!</p>
-            </Matrix>
-            <ThemeProvider theme={matrixTheme}>
-                <Container
-                    component="main"
-                    style={{
-                        width: "400px",
-                        height: "auto",
-                        position: "absolute",
-                        top: "40px",
-                        left: "25%",
+        <ThemeProvider theme={matrixTheme}>
+            <Container component="main" maxWidth="xs" color="primary">
+                <CssBaseline />
+                <Box
+                    sx={{
+                        marginTop: 8,
+                        display: "flex",
+                        flexDirection: "column",
+                        alignItems: "center",
                     }}
                 >
-                    <CssBaseline />
-                    <Box
-                        sx={{
-                            margin: "10px",
-                            bgcolor: "primary.dark",
-                        }}
-                    >
-                        <Box
-                            sx={{
-                                marginTop: 8,
-                                display: "flex",
-                                flexDirection: "column",
-                                alignItems: "center",
-                            }}
-                        >
-                            <Avatar sx={{ m: 1, bgcolor: "primary.light" }}>
-                                <LockOutlinedIcon
-                                    sx={{ color: "primary.main" }}
-                                />
-                            </Avatar>
-                            <Typography
-                                variant="h5"
-                                sx={{
-                                    mt: 5,
-                                    color: "primary.light",
-                                    fontSize: "30px",
-                                    fontWeight: 700,
-                                    fontStyle: "italic",
-                                }}
+                    <Avatar sx={{ m: 1, bgcolor: "primary.main" }}>
+                        <LockOutlinedIcon />
+                    </Avatar>
+                    <Typography component="h1" variant="h5">
+                        Issue Tracker?
+                    </Typography>
+                    <Box>
+                        <Link href={loginUri}>
+                            <Button
+                                fullWidth
+                                variant="contained"
+                                color="primary"
+                                sx={{ mt: 3 }}
                             >
-                                Issue Tracker?
-                            </Typography>
-                            <Box>
-                                <Link href={loginUri}>
-                                    <Button
-                                        sx={{
-                                            variant: "outline",
-                                            border: "1px solid",
-                                            mt: 3,
-                                            color: "primary.light",
-                                            fontSize: "12px",
-                                        }}
-                                    >
-                                        GitHub Login 깃헙 로그인
-                                    </Button>
-                                </Link>
-                            </Box>
-                        </Box>
-                        <Copyright
-                            sx={{
-                                mt: 8,
-                                mb: 4,
-                                color: "primary.light",
-                                fontSize: "10px",
-                            }}
-                        />
+                                GitHub Login 깃헙 로그인
+                            </Button>
+                        </Link>
                     </Box>
-                </Container>
-            </ThemeProvider>
-        </>
+                </Box>
+                <Copyright sx={{ mt: 8, mb: 4 }} />
+            </Container>
+        </ThemeProvider>
     );
 }
