+++ FE/tracker/src/mui-style/theme.jsx
@@ -1,0 +1,14 @@
+import { ThemeProvider, createTheme } from "@mui/material";
+
+export const matrixTheme = createTheme({
+    palette: {
+        primary: {
+            main: "#0D0208",
+        },
+        secondary: {
+            main: "#00FF41",
+            // green: "#008F11",
+            // darkgreen: "#003B00",
+        },
+    },
+});
